[REFACTOR] : ♻️ UploadedFileTab interface 통합
