--- src/components/UploadedFileTab/UploadedOne/interfaces.ts
@@ -1,15 +1,0 @@
-export interface File {
-  enrollDay: string;
-  eventId: number;
-  file: string;
-  fileName: string;
-  title: string;
-  userName: string;
-  eventType?: string;
-  userId: string;
-}
-
-export interface Props {
-  file: File;
-  type: 'myfiles' | 'meetingfiles' | 'reportfiles';
-}
